+++ ngClient/App/settings/Setup/disqus.js
@@ -1,7 +1,0 @@
-﻿(function () {
-    'use strict';
-
-    angular.module('main')
-        .constant('disqusShortname', ''); // Enter your disqus shortname here
-
-})();