+++ ngClient/App/settings/Setup/googleAnalytics.js
@@ -1,8 +1,0 @@
-﻿(function () {
-    'use strict';
-
-    angular.module('main')
-        .constant('analyticsTrackingCode', '') // Enter your google analytics tracking code here
-        .constant('analyticsDomainName', 'none'); // Enter your domain name here - 'none' for localhost
-
-})();